requirements.txt
absl-py
asgiref
astor
cached-property
certifi
click
gast
google-pasta
grpcio
gunicorn
h5py
importlib-metadata
itsdangerous
Jinja2
joblib
Keras
Keras-Applications
Keras-Preprocessing
Markdown
MarkupSafe
numpy
pandas
Pillow
protobuf
python-dateutil
pytz
PyYAML
scikit-learn
scipy
six
sqlparse
tensorboard
tensorflow
tensorflow-estimator
termcolor
threadpoolctl
typing-extensions
Werkzeug
wincertstore
wrapt
zipp
matplotlib
streamlit
streamlit-option-menu==0.3.6
opencv-python

